New additions and increased accuracy.
	There's now the functionality to use a muon beam the diameter
and shape of the decay which starts right before the hadron monitor,
instead of starting with protons on the target. Check
src/NumiDataInput.cc for the variable "useMuonBeam".
	Improved the accuracy of the Absorber Core Blocks by including
variable density blocks that outline the water cooling pipes. These
blocks roughly approximate the decreased amount of matter in these areas.
	Placed the Hadron Monitor into a mother volume that will
actually detect particles now. Had previously placed it in a volume
which overlapped another that acted as a tunnel through the Hadron
Monitor
        Changed the variables of the generic input macro
"G4Nevents.mac"to better control properties of the Monte Carlo. The
new text for running protons/muons through the G4 MC is
/NuMI/rndm/setRndmSeed 40000
/NuMI/run/setRunID 40000
/run/beamOn 40000
wherein all three variables above can be changed and the new output name
for both hadmmNtuple and nuNtuple and rndm number list is
"_40000.root" so it is easier to keep track of files.
        Added two pieces of slab steel to fill the left and right gap
between the absorber core and the surrounding blu blocks.