Changes made to accurately reflect the compsition and location of the Hadron Absorber, Absorber Hall and Muon alcoves are thoroughly chronicled in NuMI-Note 2769.
- Added the following materials to NumiMaterials
  - Shotcrete
  - Rebar_Concrete
  - Blu_Stl

- Updated the following material in NumiMaterials
  - MaqShale
  - DoloStone
  - var_Al
  - var_Stl
  - Slab_Stl (which is ASTM 836)

- Added a feature to the stepping function in NumiDataInput that
allows the user to set the kill energy threshold for
particles. Otherwise the default threshold is .05 GeV.

- Added a feature to the G4NuMI startup, where instead of a beam of
protons on the target a 'muon gun' can be used for data
generation. The muon gun is
set at the end of the decay pipe and has a random uniform distribution
across the ~2m diameter of the decay pipe. This can be turned on/off
with a boolean variable in NumiDataInput.

- Added a feature to NumiAnalysis where the input variables from
NumiDataInput (RunNumber and geometry), are added to the output file
name for easier record keeping. The RunNumber variable can be set at
runtime in by declaring the RunID i.e. /NuMI/run/setRunID.

- Corrected the Absorber Hall length and moved the first muon alcove
into the "Tunnel" physical volume in order to accomodate the
shortening of the Absorber Hall to it's more accurate surveyed
length.

- Added a feauture to the Absorber Hall, Pre-Absorber Concrete
Shielding and downstream muon alcoves(NumiHadronAbsorber and
NumiSecMonitors) that rotates and matches edges
so that the volumes are situated at the 58mrad angle of the
beam. Setting the BeamAngle variable in NumiDataInput is all the
change that's required to modify the whole geometry of the G4NuMI MC.

- Corrected a serious error in NumiMaterials where the element Calcium
had the properties of Gallium.

- Improved the accurateness of the rock composition and density that
constitutes the World
Volume. The rock is now Dolostone, a combination of Dolomite,
Maquoketa Shale and water. The whole of the beam line runs through
many different layers of rock, but the NuMI beam (up to the end of the
Muon Monitors) only passes through Dolostone.

- A feature remains that sets all irrelevant muon/hadron monitor data to
-81579, i.e. a particle hits muon monitors 1 & 2, but not three, then
the data for muon monitor 3 will be set to the nonsensical value of
-81579.