Added the 'cell' variable to the header files of the hadmmtuple data tree so that it now writes out the cell hit by the muon for each muon monitor.
Changed the G4Nevents.mac macro to reflect the addition to the detector construction. I have added the ability to change the rock and steel density as well as the muon wall location in order to simulate our best knowledge of the material between the End of the Decay Pipe and the downstream muon monitors. This is set as a funciton of sigma and so a setMaterialSigma value of 1 in G4NEvents will change all relevant values by their 1 sigma value.

The NumiRunActionMessenger now takes the cell variable when fillig the hadmmtuple, and this was not reflected in the previous header file. A cvs commit oversight.