Committing the optional (via jCompare) gnumi horn changes.
Also putting in the gnumi-like target hall shielding at M. Bishai's request.  This can be turned on and off with g3Chase (found in the same place in NumiDataInput as jCompare), but it is on by default.  Testing has shown that these blocks make approximately no difference to the neutrino flux.